fix compiler warnings, fix doc

+++ src/AudioOutput/IAudioOutput_impl.h
@@ -4,7 +4,7 @@
 
 template<typename T> void IAudioOutput::getAmplifiedBuffer(const T* inBuffer, T* outBuffer, unsigned long items)
 {
-    for(int i = 0; i<items; i++)
+    for(unsigned long i = 0; i<items; i++)
     {
         outBuffer[i] = inBuffer[i] * this->volume;
     }
